telegram_bot/requirements.txt
python-telegram-bot>=20.0
openai>=1.0.0
requests>=2.28.0
python-dotenv>=1.0.0
pydub>=0.25.0
